shows ranking in prompt
